bugfix: when you're green, start counting even if the card is still inserted.
Now if you leave the card in, when it's ok and you have credit, start counting.
If, at the end, the card is still in, loop restart from top and, if it's an old card goes red, or, if in the meaning time some has charged the card, goes green again.